+++ packages/iconv-lite/index.js
@@ -1,0 +1,11 @@
+'use strict';
+
+function identity(s) {
+  return s;
+}
+
+module.exports = {
+  decode: identity,
+  encode: identity,
+  encodingExists: () => true,
+};